# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurantes API",
        "contact": {
            "url": "",
            "email": ""
        },
        "license": {
            "name": "Deber de Programacion Web"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/usuarios": {
            "get": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Lista de usuarios",
                "operationId": "listUsuarios",
                "responses": {
                    "200": {
                        "description": "Un array de usuarios",
                        "headers": {
                            "x-next": {
                                "description": "A link to the next page of responses",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuarios"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Crear un usuario",
                "operationId": "createUsuario",
                "requestBody": {
                    "description": "Usuario a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuarios"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuarios"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuarios/{id}": {
            "get": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Obtener usuario por id",
                "operationId": "getUsuarioById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un usuario"
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "updateUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Usuario a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuarios"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuarios"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "deleteUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuarios"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error",
                "description": "Error",
                "properties": {
                    "code": {
                        "description": "Codigo de error",
                        "type": "integer",
                        "format": "int32",
                        "example": 400
                    },
                    "message": {
                        "description": "Mensaje de error",
                        "type": "string",
                        "example": "Error de validacion"
                    },
                    "errors": {
                        "description": "Errores",
                        "type": "array",
                        "items": {
                            "description": "Error",
                            "type": "string",
                            "example": "El campo nombre es requerido"
                        }
                    }
                },
                "type": "object"
            },
...
            "Usuarios": {
                "title": "Usuarios",
                "description": "Usuarios",
                "properties": {
                    "dni": {
                        "description": "DNI",
                        "type": "string",
                        "example": "12345678"
                    },
                    "nombres": {
                        "description": "Nombres",
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellidos": {
                        "description": "Apellidos",
                        "type": "string",
                        "example": "Perez"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string",
                        "example": "juanp.2@mail.com"
                    },
                    "telefono": {
                        "description": "Telefono",
                        "type": "string",
                        "example": "987654321"
                    },
                    "direccion": {
                        "description": "Direccion",
                        "type": "string",
                        "example": "Av. Los Olivos 123"
                    }
                },
                "type": "object"
            }
        }
    }
}